requirements.txt
fastapi
uvicorn
python-dotenv
openai
chromadb
langchain
pypdf
python-docx
tiktoken
pydantic
pydantic-settings
requests
